payment/src/events/listeners/order-created-listener.ts
import { Listener, OrderCreatedEvent, Subjects } from '@ventore/common';
import { queueGroupName } from './queue-group-name';
import { Message } from 'node-nats-streaming';
import { Order } from '../../models/order';

export class OrderCreatedListener extends Listener<OrderCreatedEvent> {
  queueGroupName = queueGroupName;
  subject: OrderCreatedEvent['subject'] = Subjects.OrderCreated;
  async onMessage(data: OrderCreatedEvent['data'], msg: Message) {
    const order = Order.build({
      id: data.id,
      price: data.ticket.price,
      userId: data.userId,
      status: data.status,
      __v: data.__v,
    });

    await order.save();

    msg.ack();
  }
}
